Move client plugins to client (psiphon) package
