+++ psiphon/userAgentPicker.go
@@ -17,11 +17,13 @@
  *
  */
 
-package common
+package psiphon
 
 import (
 	"net/http"
 	"sync/atomic"
+
+	"github.com/Psiphon-Labs/psiphon-tunnel-core/psiphon/common"
 )
 
 var registeredUserAgentPicker atomic.Value
@@ -54,7 +56,7 @@ func UserAgentIfUnset(h http.Header) (http.Header, bool) {
 			}
 		}
 
-		if FlipCoin() {
+		if common.FlipCoin() {
 			dialHeaders.Set("User-Agent", pickUserAgent())
 		} else {
 			dialHeaders.Set("User-Agent", "")
